Validate Source and Target of each Path
If a path was not ending / beginning on a crossing the node at the end of the last / first segment would be chosen.
We now require the segment to be the first / last one to chose a node instead of a crossing.
While the former implementation wasn't wrong the new one is clearly more restrictive.